# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.BookmarkResponseDTO:
    properties:
      createdAt:
        type: string
      id:
        type: string
      note:
        type: string
      timestampMs:
        description: 'Point 1: Already uses ms'
        type: integer
      trackId:
        type: string
      userId:
        type: string
    type: object
...
  dto.CreateBookmarkRequestDTO:
    properties:
      note:
        type: string
      timestampMs:
        description: 'Point 1: Already uses ms'
        minimum: 0
        type: integer
      trackId:
        type: string
    required:
    - timestampMs
    - trackId
    type: object
...
  dto.PaginatedProgressResponseDTO:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.PlaybackProgressResponseDTO'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  dto.PaginatedResponseDTO:
    properties:
      data:
        description: The slice of items for the current page (e.g., []AudioTrackResponseDTO)
      limit:
        description: The limit used for this page
        type: integer
      offset:
        description: The offset used for this page
        type: integer
      page:
        description: Current page number (1-based)
        type: integer
      total:
        description: Total number of items matching the query
        type: integer
      totalPages:
        description: Total number of pages
        type: integer
    type: object
  dto.PlaybackProgressResponseDTO:
    properties:
      lastListenedAt:
        type: string
      progressMs:
        description: 'Point 1: Already uses ms'
        type: integer
      trackId:
        type: string
      userId:
        type: string
    type: object
  dto.RecordProgressRequestDTO:
    properties:
      progressMs:
        description: 'Point 1: Already uses ms'
        minimum: 0
        type: integer
      trackId:
        type: string
    required:
    - progressMs
    - trackId
    type: object
...
  httputil.ErrorResponseDTO:
    properties:
      code:
        description: Application-specific error code (e.g., "INVALID_INPUT", "NOT_FOUND")
        type: string
      message:
        description: User-friendly error message
        type: string
      requestId:
        description: Include request ID for tracing
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support Team
    url: http://www.example.com/support
  description: API specification for the backend of the Language Learning Audio Player
    application. Provides endpoints for user authentication, audio content management,
    and user activity tracking.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Language Learning Audio Player API
  version: 1.0.0
paths:
...
  /bookmarks:
    get:
      description: Retrieves a paginated list of bookmarks for the authenticated user,
        optionally filtered by track ID.
      operationId: list-bookmarks
      parameters:
      - description: Filter by Audio Track UUID
        format: uuid
        in: query
        name: trackId
        type: string
      - default: 50
        description: Pagination limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of bookmarks (timestampMs in milliseconds)
          schema:
            allOf:
            - $ref: '#/definitions/dto.PaginatedResponseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.BookmarkResponseDTO'
                  type: array
              type: object
        "400":
          description: Invalid Track ID Format (if provided)
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: List user's bookmarks
      tags:
      - User Activity
    post:
      consumes:
      - application/json
      description: Creates a new bookmark at a specific timestamp (in milliseconds)
        in an audio track for the authenticated user.
      operationId: create-bookmark
      parameters:
      - description: Bookmark details (timestampMs in milliseconds)
        in: body
        name: bookmark
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookmarkRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Bookmark created successfully (timestampMs in milliseconds)
          schema:
            $ref: '#/definitions/dto.BookmarkResponseDTO'
        "400":
          description: Invalid Input / Track ID Format
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "404":
          description: Track Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Create a bookmark
      tags:
      - User Activity
  /bookmarks/{bookmarkId}:
    delete:
      description: Deletes a specific bookmark owned by the current user.
      operationId: delete-bookmark
      parameters:
      - description: Bookmark UUID
        format: uuid
        in: path
        name: bookmarkId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Bookmark deleted successfully
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "403":
          description: Forbidden (Not Owner)
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "404":
          description: Bookmark Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
      security:
      - BearerAuth // Apply the security definition defined in main.go: []
      summary: Delete a bookmark
      tags:
      - User Activity
...
  /users/me/progress:
    get:
      description: Retrieves a paginated list of playback progress records for the
        authenticated user.
      operationId: list-playback-progress
      parameters:
      - default: 50
        description: Pagination limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of playback progress (progressMs in milliseconds)
          schema:
            $ref: '#/definitions/dto.PaginatedProgressResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: List user's playback progress
      tags:
      - User Activity
    post:
      consumes:
      - application/json
      description: Records or updates the playback progress for a specific audio track
        for the authenticated user.
      operationId: record-playback-progress
      parameters:
      - description: Playback progress details (progressMs in milliseconds)
        in: body
        name: progress
        required: true
        schema:
          $ref: '#/definitions/dto.RecordProgressRequestDTO'
      produces:
      - application/json
      responses:
        "204":
          description: Progress recorded successfully
        "400":
          description: Invalid Input / Track ID Format
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "404":
          description: Track Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Record playback progress
      tags:
      - User Activity
  /users/me/progress/{trackId}:
    get:
      description: Retrieves the playback progress for a specific audio track for
        the authenticated user.
      operationId: get-playback-progress
      parameters:
      - description: Audio Track UUID
        format: uuid
        in: path
        name: trackId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Playback progress found (progressMs in milliseconds)
          schema:
            $ref: '#/definitions/dto.PlaybackProgressResponseDTO'
        "400":
          description: Invalid Track ID Format
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "404":
          description: Progress Not Found (or Track Not Found)
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Get playback progress for a track
      tags:
      - User Activity
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...